+++ LinuxPBA/UnlockSEDs.cpp
@@ -60,7 +60,7 @@ uint8_t UnlockSEDs(char * password) {
     }
     std::sort(devices.begin(),devices.end());
     printf("\nScanning....\n");
-    for(uint16_t i = 0; i < devices.size(); i++) {
+    for(size_t i = 0; i < devices.size(); i++) {
                 snprintf(devref,23,"/dev/%s",devices[i].c_str());
         tempDev = new DtaDevGeneric(devref);
         if (!tempDev->isPresent()) {
